ATTD completo con tkinter-interface
