+++ features/ahorcado.feature
@@ -1,28 +1,28 @@
-Feature: showing off behave
+Feature: Ahorcado
   Scenario: Repetir letra usada
     Given la palabra es PALABRA
     When jugador intenta las letras B,M,B
     Then se muestra pantalla usada
 
-  Scenario: Ganar el juego intentanto palabra entera
+  Scenario: Ganar el juego intentando palabra entera
     Given la palabra es PALABRA
     When jugador ingresa PALABRA
     Then se muestra pantalla ganaste
   
+  Scenario: Reiniciar juego luego de ganar
+    Given se muestra pantalla ganaste
+    When jugador clickea en reiniciar
+    Then se muestra pantalla empezar_juego
+
+  Scenario: Perder el juego intentando palabra entera
+    Given la palabra es PALABRA
+    When jugador ingresa MANZANA
+    Then se muestra pantalla perdiste
+
   Scenario: Reiniciar juego luego de perder
     Given se muestra pantalla perdiste
     When jugador clickea en reiniciar
     Then se muestra pantalla empezar_juego
-
-  Scenario: Reiniciar juego luego de ganar
-    Given se muestra pantalla ganaste
-    When jugador clickea en Reiniciar
-    Then se muestra pantalla empezar_juego
-
-  Scenario: Perder el juego intentanto palabra entera
-    Given la palabra es PALABRA
-    When jugador ingresa MANZANA
-    Then se muestra pantalla perdiste
 
   Scenario: Quedarse sin vidas
     Given la palabra es PALABRA
